xnu-7195.81.3

+++ bsd/netinet6/in6_pcb.h
@@ -91,7 +91,7 @@
  */
 
 #ifndef _NETINET6_IN6_PCB_H_
-#define	_NETINET6_IN6_PCB_H_
+#define _NETINET6_IN6_PCB_H_
 #include <sys/appleapiopts.h>
 
 #ifdef BSD_KERNEL_PRIVATE
@@ -118,7 +118,7 @@ extern int in6_getsockaddr(struct socket *, struct sockaddr **);
 extern int in6_getsockaddr_s(struct socket *, struct sockaddr_in6 *);
 extern int in6_mapped_sockaddr(struct socket *so, struct sockaddr **nam);
 extern int in6_mapped_peeraddr(struct socket *so, struct sockaddr **nam);
-extern int in6_selecthlim(struct in6pcb *, struct ifnet *);
+extern uint8_t in6_selecthlim(struct in6pcb *, struct ifnet *);
 extern int in6_pcbsetport(struct in6_addr *, struct inpcb *,
     struct proc *, int);
 extern void init_sin6(struct sockaddr_in6 *sin6, struct mbuf *m);
